Fixed a number of issues with the game
Reved to version 1.1.47 (Alpha 12), Preformed a minor refactor of the
GameController
added tactical drone counts to the screen. Fixed an issue where allies
add to the players score when
destroyed. Fixed another issue where tactical drones were not being
counted as destroyed. Decreased
the crate drop rate of the battle cruiser. Fixed some issues with how
and when battle cruisers are launched
and how the counts are handled for Survival Mode. Cruiser speed
decreases as more cruisers are launched. This makes the game
more difficult as it becomes more difficult to fight off the drones they
produce. No longer updating the screen text
when points are scored (as the tick timer updates this for us). Added
code to launch cruisers in waves after
a certain number of them have passed by. Only incrementing the score if
the objects health is at 0; this fixes
an issue where the player score increments when a battlecruiser passes
the screen. Added the game over screen overlay
this will now display "Game Over" when the player is defeated. Fixed an
issue in the projectile script where projectiles
we’re Instantiating explosions when destroyed at the end of the game.
Optimized this script so objects that detonate
only make an explosion when they call their detonate function. Modified
the menu script to enable or disable the
gyroscope depending on the setting. Added a few new game elements to
capture mode to make it operational, capture mode
has now been added to the main menu.