Tested and fixed numerous issues.
Reved to version 1.1.49, Removed the gyroscope option from the game.
Fixed an issue where the game does not reset
* the fps controller settings when the victory is achieved. Save point
now increases player health. Increased the width and boldness
* of temporary text. Adjusted the masses of each bullet to make it more
realistic. Removed code that saves game objects in the load
* game object function. Game Objects can only be saved when a save point
is passed. (this may be why game objects were doubled). Reduced
* the number of crates that get re-spawned when a crate is destroyed.
Added the new fall again script. This is to handle objects that get
* stuck in the air. Fixed an issue where grenades were not exploding.